% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDA.R
\name{qda}
\alias{qda}
\title{Discriminante Cuadratico}
\usage{
qda(x, grupo)
}
\arguments{
\item{x}{(matriz) con un marco de datos que contiene las variables predictoras}

\item{grupo}{(vector) indica a que grupo pertenece cada observacion}
}
\value{
una lista con los elementos media, covarianza, prior y log.prior.
}
\description{
Discriminante Cuadratico
}
\examples{
\dontrun{
# Crear un marco de datos con variables predictoras
x <- data.frame(
 x1 = c(1, 2, 3, 4, 5, 6),
 x2 = c(6, 5, 4, 3, 2, 1),
 x3 = c(1, 2, 1, 2, 1, 2)
 )

 # Crear un vector que indica a qué grupo pertenece cada observación
 grupo <- c('A', 'A', 'B', 'B', 'C', 'C')

 # Ejecutar la función qda_function con los datos creados
 resultado<-qda(x, grupo)

 # Ver el resultado
 print(resultado)
 }
}
